+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## [Unreleased]
+
+### Added
+- `UIConfig#seekbarAllowSeekInMarkerDuration` config property to allow seeking inside `TimelineMarker` duration
+
 ## [3.57.0] - 2024-03-28
 
 ### Added